// src/app/core/models/addCourse.ts
// dataCourse!: AddCourse => admin/course-management/addcoursemanager/addcoursemanager.component
// Observable => @service/addCourse/addcourse.service
export interface AddCourse {
  maKhoaHoc:        string;
  biDanh:           string;
  tenKhoaHoc:       string;
  moTa:             string;
  luotXem:          number;
  danhGia:          number;
  hinhAnh:          {};
  maNhom:           string;
  ngayTao:          string;
  maDanhMucKhoaHoc: string;
  taiKhoanNguoiTao: string;
}